rename: uploadDone 페이지, button 위치 수정
